Исправил недочеты в интерфейсе

+++ src/com/artem/training/store/entity/Product.java
@@ -38,8 +38,8 @@ public class Product {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
-        this.price = price;
+    public void setPrice(double price) {
+        this.price = BigDecimal.valueOf(price);
     }
 
     @Override
